jupyter
Feature scaling is a method used to normalize the range of independent variables or features of data. In data processing, it is also known as data normalization and is generally performed during the data preprocessing step.

Feature scaling helps avoid problems when some features are much larger (in absolute value) than other features.